Chantier 32.7-32.12: 14-layer deep audit of Validation/Shared/Settings/Setup/Workflow/Calendar (parallel batch 2)
Second parallel batch of the deep-audit methodology. 6 agents, each
strictly scoped to one module, merged and re-verified together (zero
file overlap except clean multi-way merges on 3 shared files —
DatabaseSeeder.php, RolesAndPermissionsSeeder.php, and
AiContextualAssistantService.php, the last touched by 2 agents at
different insertion points).

Modules/Validation: re-audited despite 3 prior passes (8.5sv/19/31) —
proof the 8-14 layers find real things the historical 7 never targeted.
Fixed 2 real IDORs (store()'s overly broad approvable_type allowlist +
no ownership check; delegateApproval() reassigning to any user
regardless of company). hierarchy_id/module_name were silently dropped
on write. Notification action_url pointed at a URL that has never
existed since Chantier 20 wired it. 'Validation' was missing from
AiContextualAssistantService and none of its 5 real pages called
useAiAssistant() — same bug class as Strategy's Chantier 30 fix.

Modules/Shared: deleted 2 dead duplicate services (real implementations
exist in Helpdesk/Analytics) + a table with zero readers/writers. Added
caching to static reference-data endpoints. Fixed a completely
unreachable web route. Flagged a severe finding: 24 jobs across
Accounting/BI extend BaseAsyncJob with no handle() method — same class
as 4 Core jobs fixed at Chantier 32.1, 6x the scale.

Modules/Settings: the module's only Vue page called zero real APIs —
pure mock state — wired for real. Fixed a real unique-constraint bug
(same key across two modules collided), a boolean-cast bug ("false"
string stored as true), and a tenant-override precedence bug. Deleted a
confirmed-dead SettingGroup model that was also independently broken.

Modules/Setup: the wizard-driven import pipeline had never successfully
imported a single row for any entity, ever — a strict_types TypeError
on tenant_id masked by zero test exercising the real dispatch path.
Fixed. Found and fixed a second, independent cross-tenant leak on the
server-rendered page (on top of the already-documented API-side fix).
Activated the fully-built AI import pipeline, which had no Vue page
calling it. Fixed field-mapping mismatches on 4 of 6 import entities.

Modules/Workflow: resolved the ApprovalWorkflowService dead-code
mandate from Chantier 31 — deleted, along with 2 newly-discovered
identical-pattern twins (TaskManagementService, WorkflowBuilderService).
Found something worse independently: WorkflowActionRegistry's
constructor took 3 params while the provider passed 18 positionally —
guaranteed TypeError on every resolution, making 12 real action-handler
module prefixes completely unreachable via live execution. Fixed.
Activated 2 real, never-routed, independently-broken controllers.

Modules/Calendar: fixed a real cross-tenant IDOR and an unauthenticated
webhook. Headline finding: 4 real Vue pages returned bare JSON arrays
from the API while every real page read response.data — these pages
have never shown real data since built. Fixed, plus 3 adjacent contract
mismatches found in the same pass.

Pest: Validation 129→151, Shared 78→53+12new, Settings 34→55,
Setup 134→150, Workflow 233+2risky→162+2risky, Calendar 46→84. Zero
regressions across 10 dependent suites (AI/Achats/Accounting/HR/
Inventory/CRM/Core/Helpdesk/Security/Strategy) + root notification
test. npm type-check/build clean.

+++ Modules/Workflow/app/Services/Actions/NotificationActionHandler.php
@@ -68,7 +68,7 @@ class NotificationActionHandler
         $to              = $context['to']               ?? $params['to']              ?? null;
         $subjectTemplate = $params['subject_template']  ?? $params['subject']         ?? 'Notification WideHalo ERP';
         $bodyTemplate    = $params['body_template']     ?? $params['body']            ?? '';
-        $tenantId        = $context['tenant_id']        ?? 1;
+        $tenantId        = (int) ($context['tenant_id']        ?? 1);
         $locale          = $params['locale']            ?? $context['locale']         ?? 'fr';
 
         if (!$to) {
@@ -160,7 +160,7 @@ class NotificationActionHandler
         $to          = $context['to']        ?? $params['to']        ?? null;
         $title       = $this->interpolate($params['title_template'] ?? $params['title'] ?? 'Notification', $context);
         $body        = $this->interpolate($params['body_template']  ?? $params['body']  ?? '', $context);
-        $tenantId    = $context['tenant_id'] ?? 1;
+        $tenantId    = (int) ($context['tenant_id'] ?? 1);
         $type        = $params['type']       ?? 'info';
         $actionUrl   = $params['action_url'] ?? null;
         $module      = $params['module']     ?? null;
@@ -240,7 +240,7 @@ class NotificationActionHandler
     {
         $to          = $context['to']        ?? $params['to']        ?? null;
         $body        = $this->interpolate($params['body_template'] ?? $params['body'] ?? '', $context);
-        $tenantId    = $context['tenant_id'] ?? 1;
+        $tenantId    = (int) ($context['tenant_id'] ?? 1);
         $gateway     = $params['gateway']    ?? $this->inferGateway($params['country'] ?? null);
 
         if (!$to || empty($body)) {
@@ -322,7 +322,7 @@ class NotificationActionHandler
         $entityId   = $context['entity_id']    ?? $context['invoice_id'] ?? null;
         $amount     = (float) ($context['amount'] ?? 0);
         $currency   = $context['currency']     ?? 'XOF';
-        $tenantId   = $context['tenant_id']    ?? 1;
+        $tenantId   = (int) ($context['tenant_id']    ?? 1);
 
         if (!$entityId) {
             return [
